Unset Force to op button pins to fix iverilog12 #1
With the Force statements on the op buttons during Reset, no test passes when running with iverilog 12. However, they do pass with iverilog 13. This is mentioned in #1. After some testing, these pins are the only ones that cannot be Force() when resetting. Investigating why. However, this is difficult due to no waveform being generated as mentioned in #2.